debugging prints for tilize, add switch back and move all classes back into cpp
