+++ ttnn/ttnn/__init__.py
@@ -94,14 +94,13 @@ def manage_config(name, value):
     logger.debug(f"Restored ttnn.CONFIG.{name} to {original_value}")
 
 
-# apparently the names need to match the types exactly for pybind function arguments, I think a pure python alias would face the same issue
 from ttnn._ttnn.multi_device import (
     MeshDevice,
     CppMeshToTensor,
-    TensorToMesh,
-    ReplicateTensorToMesh,
-    ShardTensorToMesh,
-    ShardTensorTo2dMesh,
+    CppTensorToMesh,
+    CppReplicateTensorToMesh,
+    CppShardTensorToMesh,
+    CppShardTensorTo2dMesh,
     CppConcatMeshToTensor,
     CppConcat2dMeshToTensor,
     ReplicateTensor,
