+++ ttnn/ttnn/distributed/__init__.py
@@ -6,10 +6,10 @@
 from .distributed import (
     MeshDevice,
     DispatchCoreType,
-    # TensorToMesh,
-    # ShardTensorToMesh,
-    # ShardTensor2dMesh,
-    # ReplicateTensorToMesh,
+    TensorToMesh,
+    ShardTensorToMesh,
+    ShardTensor2dMesh,
+    ReplicateTensorToMesh,
     MeshToTensor,
     ConcatMeshToTensor,
     ConcatMesh2dToTensor,
